[EI-236] [web] refactor: Refactor MoMUnitCalulator and YoYUnitCalulator to extend ChangeUnitCalculator
